kgc1222

+++ downstream/summarization/run.py
@@ -442,26 +442,27 @@ def run():
                         os.makedirs(args.output_dir+"/best_model", exist_ok=True)
                         output_model_file = os.path.join(
                             args.output_dir, "best_model/model.best.bin")
-                        # output_optim_file = os.path.join(
-                        #     args.output_dir, "best_model/optim.best.bin")
-                        # output_schedule_file = os.path.join(
-                        #     args.output_dir, "best_model/sched.best.bin")
+                        output_optim_file = os.path.join(
+                            args.output_dir, "best_model/optim.best.bin")
+                        output_schedule_file = os.path.join(
+                            args.output_dir, "best_model/sched.best.bin")
                         torch.save(model_to_save.state_dict(), output_model_file)
-                        # torch.save(optimizer.state_dict(), output_optim_file)
-                        # torch.save(scheduler.state_dict(), output_schedule_file)
+                        torch.save(optimizer.state_dict(), output_optim_file)
+                        torch.save(scheduler.state_dict(), output_schedule_file)
 
-                    logger.info(
-                        "** ** * Saving fine-tuned model and optimizer ** ** * ")
-                    model_to_save = model.module if hasattr(model, 'module') else model  # Only save the model it-self
-                    output_model_file = os.path.join(
-                        args.output_dir, "model.{0}.bin".format(i_epoch))
-                    output_optim_file = os.path.join(
-                        args.output_dir, "optim.{0}.bin".format(i_epoch))
-                    output_schedule_file = os.path.join(
-                        args.output_dir, "sched.{0}.bin".format(i_epoch))
-                    torch.save(model_to_save.state_dict(), output_model_file)
-                    torch.save(optimizer.state_dict(), output_optim_file)
-                    torch.save(scheduler.state_dict(), output_schedule_file)
+                    if i_epoch % args.keep_last_epochs == 0:
+                        logger.info(
+                            "** ** * Saving fine-tuned model and optimizer ** ** * ")
+                        model_to_save = model.module if hasattr(model, 'module') else model  # Only save the model it-self
+                        output_model_file = os.path.join(
+                            args.output_dir, "model.{0}.bin".format(i_epoch))
+                        output_optim_file = os.path.join(
+                            args.output_dir, "optim.{0}.bin".format(i_epoch))
+                        output_schedule_file = os.path.join(
+                            args.output_dir, "sched.{0}.bin".format(i_epoch))
+                        torch.save(model_to_save.state_dict(), output_model_file)
+                        torch.save(optimizer.state_dict(), output_optim_file)
+                        torch.save(scheduler.state_dict(), output_schedule_file)
 
                     # pretrained_path = os.path.join(args.output_dir, "pretrained_{0}".format(i_epoch))
                     # model_to_save.save_pretrained(pretrained_path)
@@ -472,22 +473,22 @@ def run():
                     logger.info("***** CUDA.empty_cache() *****")
                     torch.cuda.empty_cache()
 
-                    if args.keep_last_epochs > 0:
-                        # remove old epoch checkpoints; checkpoints are sorted in descending order
-                        checkpoints = checkpoint_paths(args.output_dir, pattern=r"model.\d+.bin")
-                        for old_chk in checkpoints[args.keep_last_epochs:]:
-                            if os.path.lexists(old_chk):
-                                os.remove(old_chk)
-
-                        checkpoints = checkpoint_paths(args.output_dir, pattern=r"optim.\d+.bin")
-                        for old_chk in checkpoints[args.keep_last_epochs:]:
-                            if os.path.lexists(old_chk):
-                                os.remove(old_chk)
-
-                        checkpoints = checkpoint_paths(args.output_dir, pattern=r"sched.\d+.bin")
-                        for old_chk in checkpoints[args.keep_last_epochs:]:
-                            if os.path.lexists(old_chk):
-                                os.remove(old_chk)
+                    # if args.keep_last_epochs > 0:
+                    #     # remove old epoch checkpoints; checkpoints are sorted in descending order
+                    #     checkpoints = checkpoint_paths(args.output_dir, pattern=r"model.\d+.bin")
+                    #     for old_chk in checkpoints[args.keep_last_epochs:]:
+                    #         if os.path.lexists(old_chk):
+                    #             os.remove(old_chk)
+                    #
+                    #     checkpoints = checkpoint_paths(args.output_dir, pattern=r"optim.\d+.bin")
+                    #     for old_chk in checkpoints[args.keep_last_epochs:]:
+                    #         if os.path.lexists(old_chk):
+                    #             os.remove(old_chk)
+                    #
+                    #     checkpoints = checkpoint_paths(args.output_dir, pattern=r"sched.\d+.bin")
+                    #     for old_chk in checkpoints[args.keep_last_epochs:]:
+                    #         if os.path.lexists(old_chk):
+                    #             os.remove(old_chk)
 
                 logger.info("***** CUDA.empty_cache() *****")
                 torch.cuda.empty_cache()
